backend-python/requirements.txt
Flask
mysql-connector-python
google-cloud-container
